Add "Capture On Click" support
Capture on click will most probably be infeasible on Wayland,
therefore the functionality is located in the platform-specific
part of KSG and uses native functionality (xcb/Qt5 native event
filtering).

The UI has been suitably changed to support an On-Click option,
which can be disabled on platforms where On Click will be
unavailable.

Startup times have decresed by at least 10 milliseconds (due to
the removal of a timer), probably more.