Add selection size information
KSnapshot has this very nice feature where the selection size is
visible. This made it possible to use it as a measurement tool or create
screenshots with precise dimensions so restore this feature.

The 100x100 boundary was taken from KSnapshot while the positioning
policy, colors and boundaries were chosen by me.

Tested with light and dark background preference.

REVIEW: 129109
BUG: 357080
FIXED-IN: 16.12