Fix broken previews for small selection
The use of QGraphicsDropShadowEffect in KSImageWidget enables opacity.
When the previous format is RGB32, it can be converted directly to
ARGB32 with no data changes (assuming that the pixel data really has the
format 0xffRRGGBB).

Apparently the data returned by XCB has varying values (including zero)
for the alpha channel which violates the RGB32 format and results in
fully transparant pictures. Explicitly set the bits to fix this.

REVIEW: 129135
BUG: 362450
FIXED-IN: 16.12