Show message if kipi-plugins are not installed
If the kipi-plugins package is not installed as dependency of kipi (Arch
Linux) or if it failed to load due to an ABI mismatch, the Share menu
would be empty.

While at it, move setting of mKipiMenuLoaded to avoid a race condition
when the menu is opened too fast.

REVIEW: 129168
BUG: 358557
FIXED-IN: 16.12