Fix: use autosaveFilename format also for files exported via DnD or menu
Summary:
When exporting the current screenshot directly to applications via a
temporary file holding a copy of the screenshot, the very filename of the
temporary file is then picked up in those programs, e.g. when displaying
the file or creating a local copy. Additionally that name is also passed
in the x-kde-suggestedfilename property on DnD.
With the usage of QTemporaryFile this filename holds some random part, which
makes it less useful.
Switching to a QTemporaryDir instead for getting a non-conflicting
namespace in the temp dirs, and using a normal file with the configured
autosaveFilename format in that dir results in the expected naming of
screenshot files also on export to apps via menu or DnD.

BUG: 382718
FIXED-IN: 17.08.02

Reviewers: bgupta, alexeymin

Reviewed By: alexeymin

Subscribers: alexeymin, #kde_applications, #plasma

Differential Revision: https://phabricator.kde.org/D7725