(2/3) Move "Configure..." from save menu into its own button
Summary:
Part 2 of 3 for {T7841}
Depends on D10283

BUG: 375965
FIXED-IN: 18.04

Nobody can ever find Spectacle's Configure window because the menu item to invoke is hidden in the Save menu, oddly enough. Once {D10283} lands, there'll be room for another button. So, we move the "Configure..." item out of the Save menu and into its own button.

Test Plan:
Tested in KDE Neon:
- "Save" split button no longer has a "Preferences" item
- Button works and opens the Configure Spectacle dialog
- Button responds to standard ctrl+shift+, keyboard shortcut
- Button has standard "configure" icon

{F5695269}

Reviewers: #spectacle, alexeymin, rkflx

Reviewed By: #spectacle, alexeymin, rkflx

Subscribers: alexeymin, rkflx

Differential Revision: https://phabricator.kde.org/D10289