Move printing to a new tools button
Summary:
{nav Print} does neither belong to {nav Save} nor {nav Save As}, so it
goes into a new menu button called {nav Tools}. More entries will be
added to the menu in D10295 and possibly in T6321.

In order to gain more horizontal space, {nav Export Image...} is renamed
to {nav Export}, which is in line with the other buttons not repeating
"image". "..." is removed as it should be only added to the menu entries
itself if necessary and the arrow on the button is enough to indicate
its purpose.

Ref. T7841

Test Plan:
{F5697449}

- Printing via the button still works, as well as {key Ctrl P}.
- Somewhat sensible icon chosen for the button.
- Docbook updated.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D10371