Improve usability of Save dropdown button
Summary:
The {nav Save} button allows to both {nav Save} and {nav Save As}.
Instead of showing one option on the button and one option in the menu,
we now always show both options in the menu while keeping the default
state on the button.

This makes the button easier to use, because:
- The menu always looks the same.
- It's clear which options are available and which is the default.
- Both shortcuts are displayed.
- A menu with a single item is awkward.
- Okular's selection dropdown button shows all options in the menu too.

Depends on D10467
Ref T7841

Test Plan:
Both options still work as before and the last-used button is set as the
default.

{F5705489}

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Maniphest Tasks: T7841

Differential Revision: https://phabricator.kde.org/D10468