(3/3) add option to quit after save or copy operations
Summary:
Part 3 of 3 for {T7841}

FEATURE: 389773
FIXED-IN: KDE Applications 18.04

We add a checkbox visible on the main UI to quit Spectacle after save or copy operations. This yields the following benefits:
- We can remove the "Save & Exit" item from the split button
- Users gain the ability to have spectacle quit after copying the image to the clipboard

I tried to make the feature apply to Export operations as well, but that proved to be much more complicated and I decided to abandon that for now and do it later in a subsequent patch.

Test Plan:
{F5691970}

Tested in KDE Neon:
- With "Quit after Copy or Save" unchecked:
-- Save: image is saved
-- Save As: file save dialog shown, image is saved
-- Copy to Clipboard: message is shown, image is copied to clipboard

- With "Quit after Copy or Save" checked:
-- Save: notification shown, Spectacle quits, image is saved
-- Save As: file save dialog shown, notification shown, Spectacle quits, image is saved
-- Copy to Clipboard: image is copied to clipboard (if Klipper is set up to accept images), Spectacle quits

Reviewers: #spectacle, rkflx

Reviewed By: rkflx

Subscribers: rkflx

Differential Revision: https://phabricator.kde.org/D10301