Improve opening the screenshots folder for missing config files
Summary:
If {nav Save As} was never used before and the {nav Save} button was
showing the {nav Save As} action, Spectacle would return `Pictures` by
default as the last used "Save As" location, while in all other cases it
would be `Folder/`. The missing `/` resulted in the filemanager
highlighting the folder only, instead of opening it.

The fix simply makes `lastSaveAsLocation()` consistently return URLs
with a `/` at the end.

Test Plan:
- `rm ~/.config/spectaclerc`
- Open Spectacle, {nav Tools > Open Screenshots Folder}
- Observe folder being opened instead of highlighted

Reviewers: #spectacle, gregormi, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D10559