Reorganize tools button menu for best user experience
Summary:
After 5206cdda2292, 7034b50c5c9e and 45c37a64e1cc were worked on in
parallel and added entries to the new {nav Tools} menu, it was necessary
to put them in the right order in the end.

The suggestion in D10542#207008 was met with approval, as
{nav Open Screenshots Folder} seems like an often used feature and
therefore belongs on top, and {nav Record Screen} is only slightly
related to the actual screenshot and thus can go at the bottom.

Ref T7841

Test Plan
Menu still works, same order as in planned docbook patch (see D10542).

{F5713238}

Reviewers: #spectacle, gregormi, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Maniphest Tasks: T7841

Differential Revision: https://phabricator.kde.org/D10561