Remove "Remember last used Save mode" checkbox
Summary:
Per Nate Graham: "This UI control is a relic from a prior version of spectacle when the "Save" split button's menu was full of items. Now, it only has two, and this setting is checked by default. The use case for turning it off is pretty difficult to envision now. Therefore, in the interest of de-cluttering, let's remove it."

BUG: 390809

Before:
{F5723847}

After:
{F5723849}

Test Plan:
# Select `Save` button.
# Close Spectacle.
# Open Spectacle.
# `Save` is active.
# Select `Save As`.
# Close Spectacle.
# Open Spectacle.
# `Save As` is active.

Reviewers: rkflx, ngraham, #spectacle

Reviewed By: rkflx, ngraham, #spectacle

Subscribers: acrouthamel, ngraham, rkflx

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D10711