Convert tabs to spaces in docbook
Summary:
Tabs in the docbook often cause contributor's editors to introduce
accidental whitespace changes. It was agreed to change to spaces like in
most other files.

This also corrects a problem where there were tabs inside the content.

Thanks to @gregormi for the initial version of the patch.

Test Plan: `checkXML5 index.docbook`

Reviewers: #spectacle, #documentation, ngraham, gregormi

Reviewed By: #spectacle, ngraham

Subscribers: gregormi, ltoscano, yurchor, rkflx, lueck, ngraham

Differential Revision: https://phabricator.kde.org/D10529