Update docbook screenshots after latest UI rearrangements
Summary:
Several commits attached to T7841 resulted in the screenshots being out
of date, e.g. the {nav Configure} button is not hidden anymore, there is
a new {nav Tools} menu as well as an additional option, and the default
{nav Save} option changed.

Note: With more changes planned for {nav Configure > Save}, the
corresponding screenshot was left unchanged for now.

Ref T7841

Test Plan: Compare screenshots in KHelpCenter with the real application.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Maniphest Tasks: T7841

Differential Revision: https://phabricator.kde.org/D11496