Improve on-screen help text for rectangular selection
Summary:
D11599 adds a magnifier to the rectangular region selection screen, but
the corresponding help text on how to toggle the magnifier temporarily
is a bit hidden: Most users won't discover a mouse-over help text on a
checkbox in the config dialog.

To make the feature easier to notice, the corresponding text is added to
the on-screen help. As a prerequisite, the very long line of text is
first broken up into a more tabular style.

The second help text is also broken up and reworded. Both border styles are
adapted to slightly rounded corners.

Setting the default visibility of the help texts to `false` avoids
showing them initially for a brief moment.

Test Plan:
`spectacle -r`, observe help texts with {nav Remember selected area} set
to on and also to off.

Before:
{F5828378}

After:
{F5829624}

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: guotao, ngraham

Differential Revision: https://phabricator.kde.org/D12617