Respect system color scheme for rectangular selection
Summary:
As a follow-up to 4e2f90a30b66, in addition to using the system's color
for the text labels, we now also support user-chosen colors for the
label's background and border. The highlight color for the selection
handles and the magnifier is improved, too.

The {nav Use light background} setting is now strictly about the
background colour and won't affect the selection handles anymore.

Note: Due to the opacity, we have to use `systemPalette.light` instead
of `systemPalette.window`, otherwise for light schemes the contrast
would be too low.

Test Plan:
For all color schemes, start `spectacle -r` and observe the help text in
the center and on the bottom. Look at the selection handles and press
{key Shift} to show the magnifier's crosshair. Everything should respect
the chosen color scheme and have enough contrast.

Reviewers: #spectacle, broulik, ngraham

Reviewed By: #spectacle, broulik, ngraham

Subscribers: ngraham, broulik, abalaji

Differential Revision: https://phabricator.kde.org/D12657