Deduplicate code for quitting
Summary:
By introducing `quit()`, we can have a dedicated place to handle
"quitting" the main windows by hiding it and only later exiting for
real.

We add a comment to point out that the quitting code still has issues,
which should hopefully be easier to solve with the deduplication.

Depends on D13244

Test Plan:
{nav Copy to Clipboard}, {nav Save} and {nav Save As} should work as
before with {nav Quit after Save or Copy} checked.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D13245