Deduplicate code for showing inline messages
Summary:
By adding `showInlineMessage()`, we can avoid duplicating code to
control showing and hiding of `mMessageWidget`. This also ensures
consistent use of icons and hiding timeouts.

In addition, it should be easier to add new slots for new functionality
which also needs to show inline messages, e.g. D13221.

We also introduce a way to disable automatic hiding of the inline
message, to be used in a later patch.

Depends on D13245

Test Plan:
{nav Copy to Clipboard} and {nav Export > Share > Imgur} should show
inline messages as before.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: broulik, ngraham

Differential Revision: https://phabricator.kde.org/D13247