Auto-copy shared image link to clipboard
Summary:
Whenever an image is shared and a link is provided, the link is automatically copied to the clipboard.

We also rename the {nav Copy save location} checkbox to prevent user confusion with this new feature (and also to better reflect what the {nav Copy save location} feature actually does).

FEATURE: 392722
FIXED-IN 18.08.0

Test Plan:
Shared an image via Imgur. The link immediately appeared at the top of Klipper's history.

{F5918565}

Reviewers: #spectacle, #kde_applications, rkflx

Reviewed By: #spectacle, rkflx

Subscribers: kde-doc-english, rkflx

Tags: #documentation

Differential Revision: https://phabricator.kde.org/D13493