Improve inline documentation for magnifier toggle feature
Summary:
The inline documentation for the magnifier feature is slightly confusing in a way that could frustrate users: since it's not mentioned that holding down {key Shift} only toggles the magnifier while resizing the selection rectangle, there's the very likely potential for the user to hit shift during another time, notice that nothing has happened, and conclude that the feature is broken.

This patch improves the documentation to make it clear when holding the {key Shift} will do something.

Reviewers: #spectacle

Differential Revision: https://phabricator.kde.org/D14196