Fix missing notification title found by -Wswitch-enum
Summary:
When capturing a screenshot in background mode or when
{nav Quit after Save or Copy} is checked, Spectacle shows a notfication.

Normally that notification has the capture mode set as the title, but
for {nav Window Under Cursor} it was missing (in case
{nav Capture the current pop-up only} was unchecked too).

This was due to one missing case in a `switch` statement, found by using
stricter compiler warnings.

FIXED-IN: 18.08.1

Test Plan: `spectacle -ub` should include a title in the notification again.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Differential Revision: https://phabricator.kde.org/D14716