Respect user-specified filetype for Save As... operations
Summary:
Currently, when doing {nav Save As...} operations, Spectacle outright ignores the filetype you choose in {nav Settings} and always defaults to saving as a `.png` file. This patch fixes that.

BUG: 399773
FIXED-IN: 18.12.0

Test Plan:
1. {nav Settings > Save > change the file type to something other than PNG}
2. {nav Save as...}
The file dialog now comes pre-populated with the correct file type, and saving file saves it as that file type.

Tested with all available file types.

Reviewers: #spectacle, #kde_applications, broulik, utecht, alexeymin, shubham

Reviewed By: #spectacle, alexeymin

Subscribers: aacid

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D16190