Don't save to a temp file if exporting to another app
Summary:
When exporting a screenshot to a local app for further processing, the image is saved to `/tmp/`, which causes issues in the other app, not least of which is that the file is not writable, so {nav Save} operations fail.

This patch makes Spectacle save the file to the default location when exporting to another app.

BUG: 399395
FIXED-IN: 18.12.0

Test Plan:
1. {nav Export > [some app or Other Application...]}
2. Observe that the screenshot is saved at the default location, and then opened in the app of your choice
3. Edit the image in that app and do {nav File > Save}
4. Observe that it works!

Reviewers: #spectacle, #kde_applications, broulik, utecht, alexeymin, shubham

Reviewed By: #spectacle, alexeymin

Subscribers: pino

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D16191