Add default file name token for (padded) sequential file numbering
Summary:
Default file name tokens %d will add an auto-incremented sequential file number when creating file names.
Numbering starts with 1. If the save directory has other files present with the same name (excluding
extension) but another number in the place of %d, the next number after that of the existing file will be
used to create the new file name.
The sequential number can be padded with %Nd where N is a number of total digits. If the padded token is
used in the
default file name, only file names with a number of the padded length or longer will be considered when
determining the next sequential number to use in the file name. For example, with the save name like
'Screenshot-%3d' and a save directory with only 'Screenshot-3,' the next generated file name will be
'Screenshot-001' since the existing file's number portion is not at least as long as the padded number
length.

BUG: 358641
BUG: 373759
FIXED-IN: 18.12.0

Reviewers: alexeymin, ngraham

Reviewed By: alexeymin, ngraham

Subscribers: shubham, ngraham, alexeymin, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D16304