Add option to remember rectangular region until next restart
Summary:
This adds an option to rememeber the last rectangular region until the next restart. If the user previously had selected the old remember region
checkbox the "remember across restarts will be selected" to preserve the old behavior. Also reorganised the settings page a bit. Something to
think about is changing  the "General" settings page to one specific to rectangular region as it contains only options for it.
Old:{F6621352} New:{F6621355}

BUG: 391299

Reviewers: #spectacle, #vdg, ngraham

Reviewed By: #spectacle, #vdg, ngraham

Subscribers: ngraham, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19117