Use qdbus for actions in desktop file
Summary:
As noted by @davidedmundson in D19153 Spectacle uses the spectacle command for the actions it defines in it's desktop file. This changes these
actions to use the same dbus commands as the actions defined in the khotkeys file. This opens the screenshot in the same instance of Spectacle
if it's running. Which is more expected than opening a new window if you right click on spectacle in the task manager and give it the command
"Take a new screenshot"

Test Plan:
- Open Spectacle
- Right click on it in the task manager
- No new window is opened, spectacle is now showing the new screenshot

Reviewers: #spectacle, davidedmundson, ngraham

Reviewed By: #spectacle, davidedmundson, ngraham

Subscribers: davidedmundson, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19225