[WIP] Add option to configure what happens if the Print key is pressed while Spectacle is already running
Summary:
Adds options to configure what happens if the Print key is pressed while Spectacle is already running as discussed in T9855.
Taking a new screenshot and opening a new Instance are working. If already two Instances are open only one new one will be created.
Focusing the window doesn't work it only gets marked in the panel.
Also when in a second instance the option is changed it doesn't affect the other instances altough the spectaclerc is changed.
I also need some help with the strings as currently the description is very long and not good but I couldn't think of anything suitable:
{F6623755}

Closes T9855

FEATURE: 383897
FEATURE: 374770
FEATURE: 399760
FIXED-IN: 19.04.0

Reviewers: #spectacle, #vdg, davidedmundson, ngraham

Reviewed By: #spectacle, #vdg, ngraham

Subscribers: davidedmundson, zzag, abetts, ngraham, #spectacle

Tags: #spectacle

Maniphest Tasks: T9855

Differential Revision: https://phabricator.kde.org/D19153