Fix recangle-selection help-text positioning
Summary:
BUG: 361053

When taking a screenshot with the rectangle selection tool in spectacle,
the overlaid help-text is put in the center of the entire desktop,
instead of the center of a single display. This can cause readability
issues when it is placed on the bezel of two physical displays.
This patch horizontally centers the help-text overlay on the primary display.
This affects both the "Click anywhere to start" dialogue and the hotkey help-text.

Test Plan:
Before:
{F6656417}
{F6656455}

After:
{F6656419}
{F6656456}

Reviewers: #spectacle, #vdg, ngraham

Reviewed By: #spectacle, #vdg, ngraham

Subscribers: abetts, ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19515