Add Compression Quality slider for lossy formats
Summary:
FEATURE: 63151
FIXED-IN: 19.04.0

This adds a slider to set the compression quality for lossy file formats
to the settings page, so users can set it themselves instead of having
to just accept the (in my opinion too low) default value.
The slider defaults to a quality level of 90.

A couple of notes:

  -When adding the slider to the 'Save'-page layout I noticed that the
   bottom of the help text for the filename chooser got cut off because
   the window was too small. I tried to figure out how to get the settings
   page to scroll but couldn't quite get it working. Instead I resorted to
   simply increasing the default size of the settings window just enough to make it
   fit. If someone with more experience in Qt than me can get the page to scroll OR
   can suggest a better way to display the formatting help text (which I feel
   takes up too much of the page), that would be marvelous.

  -When saving an image at full (100) quality it still looks a bit washed out and
   fuzzy, when compared to, for example, saving the image as png and converting it
   to JPG in gimp at full quality. This might be an issue with QImageWriter unless
   this is the desired behaviour? I haven't looked into it too much, just something
   I've noticed when testing

Test Plan:
This setting can be found under
Configure -> Save -> Compression Quality
{F6669261}
Set the default file type to JPG, adjust the slider and export to an image-viewer
of your choosing (eg. feh) to see how it affects image quality.

Reviewers: #vdg, #spectacle, davidre, ngraham

Reviewed By: #vdg, #spectacle, davidre, ngraham

Subscribers: ngraham, davidre, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19591