Make the filename placeholders clickable
Summary:
This allows the user to click the placeholders. First step to improve the filename template configuration. I also moved the available placeholders to the ExportManager. The SaveOptionsPage doesn't need to care about which options are present and we have to change them in only one class if we want to make changes.
Further steps I plan to do are displaying a preview, improving the Layout (we had to increase the size already) and eventually using better placeholders.

BUG: 390856
FIXED-IN: 19.04.0

Test Plan:
- Click on placeholder
- It appears in the LineEdit

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19755