Show preview of the filename in the save settings
Summary:
Shows the a preview of the configured filename template on the save settings page. I'm not sure about the placement of the label but the functionality should be there. This also again includes a small refactor to access the placeholder replacement from outside of the ExportManager.
BUG: 381175

Test Plan: {F6697579}

Reviewers: #spectacle, #vdg, ngraham

Reviewed By: #spectacle, #vdg, ngraham

Subscribers: meven, ngraham, broulik

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D19775