Respect the user-configured mimetype when dragging an image from the main window
Summary:
Currently, `ExportManager::tempSave()` expects a mimetype argument but defaults to PNG.
However no users of that function actually pass it a mimetype, so drag-and-dropped files
always get the PNG file format, even if the user changed the default file format in the
settings window.

This patch changes the function to do away with the optional mimetype argument and
instead always use the mimetype that the user has specified in the settings window.

BUG: 407165
FIXED-IN: 19.04.1

Test Plan:
- Set the file format to JPEG
- Take a new screenshot
- Drag the image from the main window to somewhere else. It is created as a JPEG file.

Reviewers: #spectacle, #kde_applications, davidre, nrother, #vdg, filipf

Reviewed By: #vdg, filipf

Subscribers: filipf

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D20976