Don't exit when running in gui mode and notification times out
This was originally commited as
https://cgit.kde.org/spectacle.git/commit/?id=c6ad1d6ae9291685b2767ab70bdeaaadb7783658,
which due to an oversight was not merged to master like it should have
been. Original commit message:

Summary: If a screenshot is taken in DBus or background mode and the running instance is 'upgraded' to gui mode while the notification is displaying the instance would still exit after the notification timed out or was destroyed. So simply check if we are running in gui mode before exiting.

Test Plan:
  # Take a screenshot via shortcut without opening Spectacle

  # While the notifcation displays open Spectacle for example via shortcut (don't run the binary directly)

  # Spectacle doesn't exit when the notifcation times out or is closed

Reviewers: #spectacle, broulik, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D21052