Show inline message after image has been saved
Summary: An inline message with the resulting filename is shown after the screenshot has been successfully saved with a button to open the containing folder. It is also possible to click on the filename to view the image.

Test Plan:
{F6842646}
Save screenshot and click on the button.The filemanager opens and the saved file is higlighted.

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D21334