Display delay in the taskmanager
Summary:
Before when taking a screenshot with delay Spectacle hid itself without any
Feedback to the user. There was no visible indication if the screenshotting
process continued, of the remaining time or if the program was still running.
This displays the display as progress in the taskmanager.

Test Plan: {F6875730}

Reviewers: #spectacle, broulik, felixernst, ngraham

Reviewed By: #spectacle, broulik, felixernst, ngraham

Subscribers: anthonyfieroni, ngraham, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D21638