Enable setting of shortcut for taking a screenshot of only the current screen
Summary:
Even though we don't set a default for this action the user can set his own
shortcut. For this to work in Spectacle we need the call to GlobalAccel as
a workaround for the same reason as the ones for the other actions.

Test Plan: Set a shortcut, should appear both in global settings and spectacle

Reviewers: #spectacle, ngraham, davidedmundson

Reviewed By: #spectacle, ngraham, davidedmundson

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D22008