Change the screenshot button to cancel during the timeout
Summary:
During the timeout period Spectacle now displays an animation in the taskmanager.
If the windows is unminized and the button clicked again (possibly multiple
times) the window will hide and show and hide again (depending on how often you
clicked) while all the button presses are processed.
To avoid this situation change the button to cancel the currently running
process. It then can also be triggered by the appropiate shortcut.

Test Plan:
- Unminimize and click the button
- Let the screenshot finish, the button is in it's original state
- cancel a region selection

Reviewers: #spectacle, broulik, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D22234