[RFC] Also animate the Cancel button
Summary:
Adds a new class ProgressButton which is a ToolButton that displays the
screenshot delay. This is achieved by first painting the button, then
a smaller transparent button on top. The width of this overlay is proportional
to the elapsed time and the color is provided by KColorScheme::PositiveText.
 Finally the icon and text are drawn on top.

Test Plan:
{F6995537}
{F6990936}
{F6990938}

Reviewers: #spectacle, broulik, mart, #vdg, ngraham

Reviewed By: #spectacle, #vdg, ngraham

Subscribers: ngraham, ndavis, #spectacle

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D22324