Port towards KConfig XT
Summary:
Port settings to Kconfig XT. This enables us to drop our own settings class and
our own configuration dialog logic with it's logic. Setting the objectNames of
the Widgets accordingly enables us to reuse them.
The settings not managed through the Dialog (last save locations and screenshot
options in the main window) are written now on exit but are still instant apply.
Using this opportunity this also unifies the naming style and moves some
settings inside the config file around: In general they are now in the same group
as they are in the settings dialog. Additionally [Save] includes
lastSave(As)Location, too. In [GuiConfig] are the options which are set in the
main window and the last crop region.
Also includes a behavioral change: "Open Screenshots Folder" will now open the
default folder as configured. The folder with the the last saved screenshot can
be opened from the inline message when saving.

BUG: 389894
FIXED-IN: 19.12.0

Test Plan: Works as before but the dialog has now more buttons

Reviewers: #spectacle, ngraham

Reviewed By: #spectacle, ngraham

Subscribers: guoyunhe, ervin, crossi, ngraham

Tags: #spectacle

Differential Revision: https://phabricator.kde.org/D23316